+++ backend/app/repositories/quotes.py
@@ -1,5 +1,6 @@
 import uuid
 
+from decimal import Decimal
 from sqlalchemy import select
 from sqlalchemy.orm import Session, joinedload, selectinload
 
@@ -79,7 +80,7 @@ class QuoteItemRepository:
         db: Session,
         quote_id: uuid.UUID,
         product_id: uuid.UUID,
-        quantity: int,
+        quantity: Decimal,
         unit_price,
     ) -> QuoteItem:
         item = QuoteItem(
